app/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flask-SocketIO
PyMySQL
Flask-CORS
python-dotenv